+++ History.md
@@ -4,6 +4,7 @@
   * Add option to generate JSCover report files (https://github.com/tntim96/JSCover/issues/215)
   * Fix undefined reportError error in jscoverage.js (https://github.com/tntim96/JSCover/issues/214)
   * Limit IE restrictions to IE7 and below
+  * Internal: Upgrade HtmlUnit 2.19
 
 1.0.22 / 2015-09-15
 ==================